resolve some todos and separate io methods
